requirements.txt
customtkinter==5.2.2
keyboard==0.13.5
Pillow==11.2.1
pynput==1.8.1
pystray==0.19.5
setproctitle==1.3.3
win32gui==221.6
